requirements.txt
appnope==0.1.3
asgiref==3.7.2
astroid==2.15.6
asttokens==2.2.1
autoflake==2.2.0
backcall==0.2.0
black==23.3.0
cfgv==3.3.1
click==8.1.3
comm==0.1.3
coverage==7.2.7
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
distlib==0.3.6
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
executing==1.2.0
filelock==3.12.2
flake8==6.0.0
gunicorn==20.1.0
identify==2.5.24
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.24.0
ipython==8.14.0
isort==5.12.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
lazy-object-proxy==1.9.0
Markdown==3.4.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nodeenv==1.8.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.1
pluggy==1.2.0
pre-commit==3.3.3
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==1.7.1
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomlkit==0.11.8
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
virtualenv==20.23.1
wcwidth==0.2.6
wrapt==1.15.0
